[Bug fixing] Fixed vector toString x, y from double to int

+++ src/main/java/org/ntut/posd2024f/shapes/TwoDimensionalVector.java
@@ -39,6 +39,6 @@ public class TwoDimensionalVector {
 
     @Override
     public String toString() {
-        return "[" + x + ", " + y + "]";
+        return "[" + (int)x + ", " + (int)y + "]";
     }
 }
